backend/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==2.2.16
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.5.0
filetype==1.2.0
gunicorn==20.0.4
idna==3.4
iniconfig==2.0.0
isort
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2020.1
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.3.1
toml==0.10.2
uritemplate==4.1.1
urllib3==2.0.2
python-dotenv
django-filter==2.4.0